[FOLDER ORGANIZATION] - Renommer aussi les fichiers dans le dossier "/console/fake" en "agendaFake.js" et "dicoFake.js" et ainsi de suite.
