+++ server/console/fake/explicationFake.js
@@ -1,3 +1,5 @@
+/* jshint node: true */
+
 var phony = require('phony').make_phony();
 var _ = require('underscore');
 var ObjectId = require('mongodb').ObjectID;
