+++ server/console/fake/newFake.js
@@ -1,43 +1,44 @@
+/* jshint node: true */
+
 var phony = require('phony').make_phony();
 var _ = require('underscore');
-var ObjectId = require('mongodb').ObjectID;
 var Repository = require('../../model/repository.js');
 var helper = require('./helper.js');
 
-var treatiesRepository = new Repository('treaties');
-var tagsRepository = new Repository('tags');
+var newsRepository = new Repository('news');
+var sourcesRepository = new Repository('sources');
 
 var days = _.sample(_.range(0, 30), _.random(15, 25));
 var months = _.range(0, 4);
 var year = 2015;
-var treaties = [];
-var allTags = [];
+var allSources = [];
+var news = [];
 
-tagsRepository.find({
-	Type: 'Treaty'
-}).then(function(docTags) {
-	allTags = docTags.map(function(t) {
-		return t._id.toString();
-	});
+sourcesRepository.find({}).then(function(sourcesDoc) {
+	allSources = sourcesDoc;
 	months.forEach(function(month) {
 		days.forEach(function(day) {
 			var nbOfItemsInDay = _.range(0, _.random(1, 4));
 			nbOfItemsInDay.forEach(function() {
-				var tags = _.sample(allTags, _.random(1, allTags.length));
-				var treaty = {
+				var citations = [];
+				for (var i = 0; i < _.random(1, 6); i++) {
+					citations.push(phony.lorem_paragraph());
+				}
+				var newOne = {
 					CreatedBy: 'parany',
 					CreatedOn: new Date().getTime(),
-					UpdatedBy: null,
+					UpdatedBy: 'parany',
 					UpdatedOn: null,
 					Date: helper.getRandomTime(year, month, day),
 					Title: phony.lorem_title(),
-					Text: helper.getText(),
-					TagIdList: tags
+					Content: helper.getText(),
+					Source: _.sample(allSources),
+					Citations: citations
 				};
-				treaties.push(treaty);
+				news.push(newOne);
 			});
 		});
 	});
-	console.log(treaties[0]);
-	treatiesRepository.save(treaties);
+	console.log(news[0]);
+	newsRepository.save(news);
 });
