+++ server/console/fake/tagFake.js
@@ -1,3 +1,5 @@
+/* jshint node: true */
+
 var phony = require('phony').make_phony();
 var _ = require('underscore');
 var Repository = require('../../model/repository.js');
