+++ server/console/fake/treatyFake.js
@@ -1,42 +1,45 @@
+/* jshint node: true */
+
 var phony = require('phony').make_phony();
 var _ = require('underscore');
+var ObjectId = require('mongodb').ObjectID;
 var Repository = require('../../model/repository.js');
 var helper = require('./helper.js');
 
-var newsRepository = new Repository('news');
-var sourcesRepository = new Repository('sources');
+var treatiesRepository = new Repository('treaties');
+var tagsRepository = new Repository('tags');
 
 var days = _.sample(_.range(0, 30), _.random(15, 25));
 var months = _.range(0, 4);
 var year = 2015;
-var allSources = [];
-var news = [];
+var treaties = [];
+var allTags = [];
 
-sourcesRepository.find({}).then(function(sourcesDoc) {
-	allSources = sourcesDoc;
+tagsRepository.find({
+	Type: 'Treaty'
+}).then(function(docTags) {
+	allTags = docTags.map(function(t) {
+		return t._id.toString();
+	});
 	months.forEach(function(month) {
 		days.forEach(function(day) {
 			var nbOfItemsInDay = _.range(0, _.random(1, 4));
 			nbOfItemsInDay.forEach(function() {
-				var citations = [];
-				for (var i = 0; i < _.random(1, 6); i++) {
-					citations.push(phony.lorem_paragraph());
-				}
-				var newOne = {
+				var tags = _.sample(allTags, _.random(1, allTags.length));
+				var treaty = {
 					CreatedBy: 'parany',
 					CreatedOn: new Date().getTime(),
-					UpdatedBy: 'parany',
+					UpdatedBy: null,
 					UpdatedOn: null,
 					Date: helper.getRandomTime(year, month, day),
 					Title: phony.lorem_title(),
-					Content: helper.getText(),
-					Source: _.sample(allSources),
-					Citations: citations
+					Text: helper.getText(),
+					TagIdList: tags
 				};
-				news.push(newOne);
+				treaties.push(treaty);
 			});
 		});
 	});
-	console.log(news[0]);
-	newsRepository.save(news);
+	console.log(treaties[0]);
+	treatiesRepository.save(treaties);
 });
